Revert "Merge pull request #39 from Revive-Graduation-Project/feature/monetary-precision-fix"
This reverts commit 6357ae5ecab2fdcad9d84213e94a5fae3e9089b3, reversing
changes made to a3cce7e67a183d0cdef63c1b513cf531c37fc481.

+++ order-service/src/main/java/com/restaurant/order/dto/snapshot/MealPriceSnapshot.java
@@ -9,6 +9,6 @@ import java.math.BigDecimal;
 public record MealPriceSnapshot(
         Long id,
         String name,
-        Double price,
+        BigDecimal price,
         String imageUrl
 ) {}
